Backend/requirements.prod.txt
# Core dependencies
bcrypt==4.0.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.7.0
gunicorn==21.2.0
itsdangerous==2.0.1
python-dotenv==0.21.1
SQLAlchemy==2.0.12
Werkzeug==2.2.2
WTForms==3.0.1